Resume mocking SAM.gov call (#2360)
We meant to mock this and when we changed to using requests.session
we missed updating the mock call in api/test_views.py.
This wasn't detected because the actual call to SAM.gov would
work—until this week, when the larger volume of SAM.gov API calls
from actual users as well as our tests pushed us past the rate limit.

Co-authored-by: Tadhg O'Higgins <2626258+tadhg-ohiggins@users.noreply.github.com>