Approach for changing old-format report_ids to new format (#2376) (#2397)
* Approach for changing old-format report_ids to new format.

* Change data_source.

* Configure Terraform's s3 backend endpoint with a URL

The old parameter has been deprecated, and there appears to be stronger requirements for a URL, not just a hostname, for the endpoint value.

* Second try at giving the Terraform S3 backend a URL endpoint

The newer argument referred to in the deprecation warning (`endpoints.s3`) is not being accepted. So we'll go back to just specifying `endpoint`, but give it a URL instead of just a hostname for the endpoint.

* Propagate S3 backend endpoint URL change to local config

This file populates the necessary Terraform config for local development (an alternative to the GHA workflow, which supplies the backend vars as secrets). This change accounts for the stricter checking on the `endpoint`, which should be a URL rather than a hostname; see the previous commit for more context.

* Provide URL for s3 backend endpoint during Terraform apply

Similar to previous commits: We have to provide a URL here, since Terraform got stricter about the value it wants for "endpoint". This commit is for the workflow used during the actual deployment.

---------

Co-authored-by: Tadhg O'Higgins <2626258+tadhg-ohiggins@users.noreply.github.com>
Co-authored-by: Bret Mogilefsky <bret.mogilefsky@gsa.gov>