Passes all regressions, demonstrates new checks
This commit now passes all workbooks through the IR.

The README is started, but not complete.

It demonstrates the authoring of checks in keeping with the
cross-validations. The errors generated pass through to the frontend
correctly, and visualize like all other workbook upload errors.

A no-op check is provided. It always passes.

A check for a missing UEI is provided. It correctly stops empty UEIs. If
a UEI is present, but does not match the regex, it passes through the
IR check, and is caught by the schema validation.