Fixes state/other cluster errors
This takes a recent helpdesk ticket and demonstrates how to improve the
error messages.

In this case, a user had

N/A N/A N/A

for cluster name, state cluster name, and other cluster name.

This causes horrible JSON Schema validation errors.

Now, we tell them exactly what they need to do in order to fix the
workbook.

Once the instructions are followed, the user's workbook is fixed, and it
passes JSON Schema validation.: