Adds transforms
The first transform is on Notes to SEFA.

We have an invisible column called seq_number. Users somehow manage to
delete things in a way that the cell ends up with a `REF!` instead of a
number. The solution is to replace the column with values that are
generated computationally on the server side.

It is not clear that we use this value?