Using section names, adding prechecks
Broke the general checks out, so they can run before transforms.

Some transforms may have to run before. (I found one in notes-to-sefa.)

So, we need to make sure we're in the right workbook. Then we can clean
up the data. Then we can begin checking it for semantics. Then we can
rewrite it into a JSON doc.

So.

Those changes are in this commit, and some tightening on the passthrough
names. All inspired by the same workbook...