Adding a failure test
This walks workbooks that are constructed to fail.

The test runs all of them, and counts the failures.

It should be the case that every workbook in the directory fails.
Therefore, we count the workbooks and count the failures.

If they come out different, clearly, we did not fail everywhere we
expected.