Fixes logic in check_federal_program_total
The check was being applied against the federal_program_total line.

However, it was summing all expenditures and expecting to find the sum
of everything in `amount_expended` to appear in `federal_program_total`.

What the check wanted to do was to compare against
`total_federal_expenditures.` I suspect this is because I did not ticket
the issue clearly.

This makes the check do the right thing.