Fixes my fix.
I was confused.

I forgot that an outcome from earlier was that we needed to fix the
workbook generator.

Therefore, I confused myself on the check_federal_program_total code.

I reverted my adventure there (leaving it as-is from HSMD's work), and
instead fixed the workbook generator to now create/insert the `cfda_key`
column.

I regenerated 100010 to verify that this now works. And, the end-to-end
code now works as well (again), because it does the right thing.

test_commands is a casualty. It is testing a fixture command that is
aging, and I think that end-to-end and the workbook generator will
replace it.