Addes auditor cert, fixes E2E regression
The text on the table changed, so the regression had to change. I've
switched to a regex instead of the full contents of the table caption.

This also makes the E2E workbook test code actually move audits into the
"Accepted" table.

This preps the E2E code for use in actuall regression testing.