Jadudm fix workbook generator for new checks (#2563)
* Fixes logic in check_federal_program_total

The check was being applied against the federal_program_total line.

However, it was summing all expenditures and expecting to find the sum
of everything in `amount_expended` to appear in `federal_program_total`.

What the check wanted to do was to compare against
`total_federal_expenditures.` I suspect this is because I did not ticket
the issue clearly.

This makes the check do the right thing.

* Fixes my fix.

I was confused.

I forgot that an outcome from earlier was that we needed to fix the
workbook generator.

Therefore, I confused myself on the check_federal_program_total code.

I reverted my adventure there (leaving it as-is from HSMD's work), and
instead fixed the workbook generator to now create/insert the `cfda_key`
column.

I regenerated 100010 to verify that this now works. And, the end-to-end
code now works as well (again), because it does the right thing.

test_commands is a casualty. It is testing a fixture command that is
aging, and I think that end-to-end and the workbook generator will
replace it.

* That was too easy.

I spent too much time with date formats.

* Linting.

* Before main sync

* Addes auditor cert, fixes E2E regression

The text on the table changed, so the regression had to change. I've
switched to a regex instead of the full contents of the table caption.

This also makes the E2E workbook test code actually move audits into the
"Accepted" table.

This preps the E2E code for use in actuall regression testing.

* Linting.

* Apply suggestions from code review

Accepting cleanup suggestions.

Co-authored-by: Phil Dominguez <142051477+phildominguez-gsa@users.noreply.github.com>

* Removing a test.

---------

Co-authored-by: Phil Dominguez <142051477+phildominguez-gsa@users.noreply.github.com>