Alternative suggestion for loading data from S3
This offers two things:

1. A small bash script for loading a directory of CSVs into Minio
   (for local dev)
2. A Django command that uses Pandas to chunk large CSV loads into
   Postgres.

The README explains more.

Why?

I was very concerned about storing data in the tree. This is not good
practice. And, it won't scale to testing larger datasets locally. Given
that we have to scale to 10-15M rows, we need to be ready to test that
on local machines. This provides a way to quickly load a set of CSVs
into Minio, and then a Django command that chunks the load of those
files into Postgres.

This can be reviewed by the data group for consideration.