Seems to run... backs up between buckets
Uses a Python-based approach and a Django command as opposed to a shell
script.

Lots of issues with configuration; this would want to be
refactored/improved after we improved how configuration for the app as a
whole takes place.

It should be much easier to access buckets, etc. regardless of the
environment we are working in (local, prod).