Moving everything to UUIDs.
This moves both admin access and tribal data access to a table that is
explicitly created via INSERTs that are static/loaded at creation-time.

The intent (which could be enforced by schema/privileges) would be to
have a table that cannot be modified by the Postgrest user. This way, we
have a table that is lookup-only. This would require creating a schema
for the table, and prohibiting anything but SELECT perms on that schema
from the api_fac_gov user.

For now, this is still "for discussion." I have tests to add, etc. This
looks very promising, however.

The admin functions "work." The file "test.rest" in
support/api/admin_api_v1_0_0 allow some quick experimentation. They'll
want to be turned into Python functions that run locally/in GH Actions
as part of a test suite.