Jadudm/fix node (#3100)
* Restructured dockerfile

This lets us do

docker build --progress=plain --target node .

in order to (say) test a build up to a given point.

It also gets rid of all of the redundant `apt update` calls. You only
need to do it once in an image.

In theory, I think we could speed up builds by taking most of the image
and publishing it via our GHCR, and *only* building the app portion at
the very end on a stock app image.

* For another test in preview.

* Changing version of npm.