Use only disseminated data in the public API (#3381)
* Fixes gross error in API.

The `OR` conjunction in every view was causing a combinatoric explosion
on every single API query, destroying performance as the DB grew.

This fixes that problem.

* Fixes the general view

Uses only disseminated data instead of the  internal table.

THis broke with historical data, because not all environments have a
historical singleauditchecklist (at this time).

* Fixing accidental removal

Leave the additional_ueis there.