Adds materialized views to the startup
Uses the run.sh refactored form to add materialized views.

These are dropped first, then created on every deploy. This is because
we might choose to change the view between deploys. In that case, we
should completely destroy it.

The Django command

fac materialized_views --refresh

can be used to refresh the view(s) on any cycle desired.