Remove exception/error on named functions
We have discovered that some accounting packages add named functions to
the defined names in a workbook.

These crash our code, because we assumed all named things were ranges.
Named functions do not (for example) have cell references.

This change leaves a logging statement in place for this situation. We
no longer report an error to the user. Why? Because the user may not
know the functions even exist.

We will still reject workbooks that have named ranges we cannot
recognize. We still only pull data from named ranges we recognize. But,
we will allow the named functions to exist.