Adds auditing for tables.
This is based on open code from Supabase. Minor changes like "IF EXISTS"
were needed, but this could be a path to providing easy auditing on
tables.

Specifically, it would let us use PostgRESTs read/write/update features
on a view to make a whole table editable via API. However, the audit log
would give us complete "before/after" tracking of the contents of the
change.