Adds new views to admin
This lets us see a lot more of the DB from an admin perspective.

Further, it adds the cognizant tables to the change tracking. These are
tables that we need to edit, and in doing so, we should then have an
audit log. We'll perhaps want to add API RPC endpoints to update these
tables, too.

For now, all it does is expose the tables via the admin API, and makes
sure that auditing gets enabled. Later, we can deal with how we track
changes.