Certification Signature - read-only date input (#4196)
For the signature date fields, I made them read-only and supplied a [select a date] placeholder. This will prompt the auditor/auditee to use the datepicker component instead of attempting to type in a date and risk invalid input.

I also updated the Cypress full-submission test where the value is normally typed in for the input. Now, Cypress will select the calendar element and select today's date.