Fix inconsistent case where datepicker fails to click (#4237)
Sometimes the full-submission tests fail on this line - it seems that the datepicker component does not always respond to .click(), and causes a timeout when the next step attempts to select today's date.

To my understanding it seems `.trigger('click')` is another approach that fares better in cases where `.click()` might fail. I have tested this many times and have yet for Cypress to break on my end with this new line.