Refactor changes for leap year calculation
Found a library `dateutil` that fulfills the same purpose as my utility method. I removed this utility and am using relativedelta() in its place.

I also have rewired the fiscal end date in `test_sac_creation.py` so that the date is based on the helpdesk ticket that created this issue. The end date is `6/30/2022` and so the expected start date is `7/1/2021`.