Consider leap years when generating fiscal period start date (#4239)
* Account for leap year when calculating fiscal start date

- Using factory calendar library for determining leap year, and then subtracting either 366 or 365 days from the fiscal end date accordingly.
- Added similar logic to general information test: `test_when_auditee_fiscal_period_end_is_valid`.

* Add Leap Year Utility

Added a new method in Util that checks the year for a leap year. This method is now referenced in `general_information` and the test.

* Sanity checks

Handling more edge cases. This is the expected behavior as of this commit:

- If the end date is `10/1/2020`, the start date will be `10/1/2019` (366 days)
- If the end date is `10/1/2021`, the start date will be `10/1/2020` (365 days)
- If the end date is `2/1/2020`, the start date will be `2/1/2019` (365 days)
- If the end date is `2/1/2021`, the start date will be `2/1/2020` (366 days)

* Refactor changes for leap year calculation

Found a library `dateutil` that fulfills the same purpose as my utility method. I removed this utility and am using relativedelta() in its place.

I also have rewired the fiscal end date in `test_sac_creation.py` so that the date is based on the helpdesk ticket that created this issue. The end date is `6/30/2022` and so the expected start date is `7/1/2021`.

* Linting fix

* Linting fix - flake8

* More linting fixes

Accounted for dateutil in `requirements.txt`.

* Sanity check for linting

Added `python-dateutil` to requirements.in.

* Update linting.yml

Running --install-types to see if this passes recent lint failures.

* Yet another sanity check on linting

* Update linting.yml

* Document changes

Recording why this change was made, how it worked previously, and noting that this change is being made after the census historical migrations have been completed.