Convert new Admin API emails to lowercase (#4225)
* Update create_functions.sql

This SQL script either gets or inserts a user for the admin API depending on a few parameters. This proposed change ensures new users have their email reduced to lowercase, and the prior verification transforms both the table and the parameter when checking if the user exists.

* Cleaning edge cases

Converting all email checks and "insertions" to lowercase when doing the following:
- Creating a new admin in the Admin API.
- Retrieving an existing admin in the Admin API.
- Removing Tribal API key access for a specified email.
- Adding read access to the Tribal API for a specified email.
- Removing Tribal access emails.
- Adding Tribal access emails.

* Admin API local testing documentation

Added a first draft of how to start using the admin API for local dev to `/docs/testing.md` .