Fix submission race condition (#4292)
* Working on a race hazard fix

This does two things:

1. Makes the dissemination fully atomic (both the data dissemination and
   state machine update)
2. Removes the "SUBMITTED" status, which was never user visible, and
   seems to play no role (other than to be confusing).

* Adding a compose file that supports two FAC instances

* Includes better comments for the race hazard

Makes it easier to communicate the exploration.

I believe this code represents how we should structure the submission to
protect against a race from multiple button presses/multiple FAC
instances at time of submission.

* Ready for testing, linting.

This cleans up the comments and code, along with a README to support the
race hazard debugging. We may yet discover more issues in this part of
the code later, so the documentation is thorough for the sake of future
interpretation of the thinking involved.

* Linting.

* Forgot to add the README.

This supports the race condition debugging.