Hotfix - continuation of #4328 (#4329)
* Update drop_schema.sql

This should drop a now deprecated `admin_api_v1_0_0`, which appears to still exist in the dev environment. It is likely to also still exist in staging and prod. This update drops the table to prevent any hassles with newer table changes, migrations, etc.

* Update drop_schema.sql

One more for the road.  Follow-up to #4328.