Simplified.
Also, added partitions.

Partitions are 3x faster for downloading all the data (by year), but 10x
slower than batches (via EXPLAIN). In the real, it is 2x slower to
download than batches, but 3x faster than a straight download.

So. Probably worth it, if we can document that people downloading all
the data should either 1) do so year-by-year, or 2) use batches.