Incremental on removal of Tribal/admin API.
This was work-in-flight. Instead of pushing further, I'm committing the in-between state. This moved some Admin API stuff around, which I'll remove in the next commit.

The config change is for local dev only. We had a bucket called `gsa-fac-private-s3`, which was legacy naming that has been superceded in all other environments. The bucket is now called `fac-private-s3`, mirroring the production environments. This should not impact any operations except for local/Docker execution. (And, it should be transparent)

The next commit should clean this up further.