Removes all spaces from prior reference numbers (#4846)
* Removes spaces from prior reference numbers

The code previously removed spaces before the comma; we want to remove
spaces entirely.

That is, regardless of whether we have

2023-001 ,2022-004

or

2023-001, 2022-004

or

2023-001 , 2022-004

they should all become

2023-001,2022-004

We already trim() the strings; this makes sure there are no internal
spaces whatsoever.

* Adding unit test

No spaces should remain in the `prior_references` field after this
transform is complete.

* Removing unneeded import

* Need to explicitly ignore a Flake8 warning

We're trying to test the space at the end of a line...

* Remove a whitespace?

* Does this pass the linter?

I really want the whitespace...