Trimming prior reference strings (#4848)
* Trimming prior reference strings

This is a partner to a previous fix.

When we pull prior reference strings, they might look like:

* "2022-001"
* "2022-001,2021-003"
* "2022-001, 2021-003"
* "2022-001 ,2021-003"
* " 2022-001,2021-003"
* ...

We were splitting on a comma, but we could then be left with substrings
that have whitespace at the beginning and/or end.

This adds a `map()` that does a `trim()` on every string. This
guarantees that there is no whitespace in the prior reference strings,
and should eliminate some lookup/validation errors.

* `strip()`, not `trim()`

---------

Co-authored-by: James Person <jperson1@umbc.edu>
Co-authored-by: James Person <james.person@gsa.gov>