Updated for full cycle
This should allow us to take raw(ish) data from prod, clean it up
further, and prepare it for reuse.

More work would be needed to work directly with prod data. We
(hopefully) will not have a need to re-run this process again. If we do,
we will extend the script "further back" so that it is easy to
replicate.