Jadudm/automatic uei waiver (#4872)
* Initial UEI waiver automation exploration

Per ADR.

* Checking migrations... interim commit.

* Adds validation waiver

I really don't know what we want to put for the name of the entity. It
might be better to leave blank. I'm going to move this to a draft PR for
others to put eyes on.

* Minor email address edit

* Adding ADR and tests

* Improving test

Adding a check to see the waiver was granted/generated.

* Adding shape validation into UEI checkups

Because we allow a 403/404, we need to make sure we match the schema
here.

* Fixing tests for new validation process

* Fixing linting.

* Update ADR to fit format

* Only grant waivers if none exist

Add logging as well.

* Fixed test case with multiple submissions